tweaks

+++ react-vite/src/utils.js
@@ -44,7 +44,7 @@ const getDayKey = (dayIndex) => {
 
 export const getDate = (date) => {
   const newDate = new Date(date);
-  const month = newDate.toLocaleString('default', { month: 'long' });
+  const month = newDate.toLocaleString('default', { month: 'short' });
   const day = newDate.getDate();
   const year = newDate.getFullYear();
   return [month, ' ', day, ', ', year]
